backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.127.1",
    "google-auth>=2.45.0",
    "passlib[bcrypt]>=1.7.4",
    
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "uvicorn>=0.40.0",
]
